Build setup
Set up the build system for multiple cpp files
with includes in a separate folder. The makefile
needs more work. This current one requires adding every
new file to the makefile in order to build it. There is a
better way to add all the files in a directory with a given
ending, but it seems to break the rest of my makefile. The
object and dependency files seem to be in the wrong place
when I use the wildcard.